% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppHungarian.R
\docType{package}
\name{RcppHungarian_package}
\alias{RcppHungarian_package}
\alias{RcppHungarian_package-package}
\title{RcppHungarian}
\description{
Header Library and R Functions to Solve Minimum Cost Bipartite Matching Problem 
using Huhn-Munkres algoritym (Hungarian Algorithym). This is a repackaging of code 
written by Cong Ma in the GitHub repo mcximing/hungarian-algorithm-cpp.
}
