% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppHungarian.R
\docType{package}
\name{RcppHungarian_package}
\alias{RcppHungarian_package}
\alias{RcppHungarian_package-package}
\title{RcppHungarian}
\description{
Header Library and R Functions to Solve Minimum Cost Bipartite Matching 
Problem using Huhn-Munkres algorithm (Hungarian algorithm; 
<https://en.wikipedia.org/wiki/Hungarian_algorithm>;
Kuhn (1955) doi:10.1002/nav.3800020109).
This is a repackaging of code written by Cong Ma in the GitHub 
repo <https://github.com/mcximing/hungarian-algorithm-cpp>.
}
